Finish sign in and sign up pages
Add CSS file to used by both pages because content is similar
enough to share the file.

The authSlice reducer now only changes state. Firebase calls have
been moved to the functions in user-api.js.

On an error with signing in/up an error message will appear below
the sign in/up form.